Fix compilation error
